♻️ Split up e2e/fixtures.ts
