+++ tests/fixtures.ts
@@ -1,181 +1,28 @@
-import { mkdir, rm } from 'node:fs/promises';
 import type { TempFilesFixture } from './fixtures/tempfiles.js';
-import type { GalleryCardSpecifier, NavigationTab, PredownloadedModel } from './utils/index.js';
-import type { Locator } from '@playwright/test';
-import type { MetadataValue, Settings } from '$lib/database';
-import type { IDBDatabaseType } from '$lib/idb.svelte';
-import type { RuntimeValue } from '$lib/schemas/metadata';
-import type { Toast } from '$lib/toasts.svelte.js';
+import type { PredownloadedModel } from './utils/index.js';
 
 import 'urlpattern-polyfill';
 
+import type { AppFixture } from './fixtures/app.js';
 import type { ONNXModelsFixture } from './fixtures/onnxmodels.js';
 
 import { defineNetworkFixture } from '@msw/playwright';
-import { test as base, expect as baseExpect } from '@playwright/test';
-import { ms } from 'convert';
-
-import { safeJSONParse } from '$lib/utils';
+import { test as base } from '@playwright/test';
 
 import _fullProtocol from '../examples/arthropods.cigaleprotocol.json' with { type: 'json' };
 import lightProtocol from '../examples/arthropods.light.cigaleprotocol.json' with { type: 'json' };
 import { FixturePaths } from './filepaths.js';
+import { app } from './fixtures/app.js';
+import { forEachTest, forEachWorker } from './fixtures/foreachtest.js';
 import { onnxmodels } from './fixtures/onnxmodels.js';
 import { tempfiles } from './fixtures/tempfiles.js';
-import {
-	confirmDeletionModal,
-	DontWait,
-	dumpDatabase,
-	expectTooltipContent,
-	galleryCard,
-	getDatabaseRowByField,
-	getDatabaseRowById,
-	getPredownloadedModel,
-	getSettings,
-	getTab,
-	goToTab,
-	listTable,
-	metadataSections,
-	mockPredownloadedModels,
-	mockProtocolSourceURL,
-	mockUrl,
-	modal,
-	setHardwareConcurrency,
-	setSettings,
-	toast,
-	waitForLoadingEnd,
-	waitForRoute,
-} from './utils/index.js';
+import { getPredownloadedModel } from './utils/index.js';
 
 const fullProtocol = _fullProtocol;
-
-type RemoveNamespace<Key extends `io.github.cigaleapp.arthropods.example.light__${string}`> =
-	Key extends `io.github.cigaleapp.arthropods.example.light__${infer Rest}` ? Rest : never;
 
 let arthropodaClassifierModel: PredownloadedModel | null = null;
 let collembolaClassifierModel: PredownloadedModel | null = null;
 let arthropodaDetectionModel: PredownloadedModel | null = null;
-
-export type AppFixture = {
-	wait: (ms: number | `${number}${'ms' | 's'}`) => Promise<void>;
-	metadata: ReturnType<typeof metadataSections>;
-	db: {
-		ready(): Promise<void>;
-		refresh(): Promise<void>;
-		list<T extends keyof IDBDatabaseType>(name: T): Promise<IDBDatabaseType[T]['value'][]>;
-		get<T extends keyof IDBDatabaseType>(
-			name: T,
-			key: string
-		): Promise<undefined | IDBDatabaseType[T]['value']>;
-		count(name: keyof IDBDatabaseType): Promise<number>;
-		protocol: {
-			byId(id: string): Promise<IDBDatabaseType['Protocol']['value'] | undefined>;
-			byName(name: string): Promise<IDBDatabaseType['Protocol']['value'] | undefined>;
-		};
-		observation: {
-			byLabel(label: string): Promise<IDBDatabaseType['Observation']['value'] | undefined>;
-			byId(id: string): Promise<IDBDatabaseType['Observation']['value'] | undefined>;
-			list(): Promise<IDBDatabaseType['Observation']['value'][]>;
-		};
-		image: {
-			byFilename(
-				filename: FixturePaths.Photos | (string & {})
-			): Promise<IDBDatabaseType['Image']['value'] | undefined>;
-			byId(id: string): Promise<IDBDatabaseType['Image']['value'] | undefined>;
-			list(): Promise<IDBDatabaseType['Image']['value'][]>;
-		};
-		session: {
-			byName(label: string): Promise<IDBDatabaseType['Session']['value'] | undefined>;
-			byId(id: string): Promise<IDBDatabaseType['Session']['value'] | undefined>;
-			list(): Promise<IDBDatabaseType['Session']['value'][]>;
-		};
-		metadata: {
-			get(id: string): Promise<IDBDatabaseType['Metadata']['value'] | undefined>;
-			of(args: {
-				/** The image's filename */
-				image?: string;
-				/** The image's ID */
-				imageId?: string;
-				/** The observation's label */
-				observation?: string;
-				/** The session's name */
-				session?: string;
-				/** Remove namespace from metadata id (keys of returned object). By default, set to lightweight protocol's id */
-				protocolId?: string | null;
-			}): Promise<
-				Record<
-					string,
-					Omit<MetadataValue, 'value'> & { rawValue: string; parsedValue: RuntimeValue }
-				>
-			>;
-			values(args: {
-				/** The image's filename */
-				image?: string;
-				/** The image's ID */
-				imageId?: string;
-				/** The observation's label */
-				observation?: string;
-				/** The session's name */
-				session?: string;
-				/** Remove namespace from metadata id (keys of returned object). By default, set to lightweight protocol's id */
-				protocolId?: string | null;
-			}): Promise<Record<string, RuntimeValue>>;
-			set(
-				/** The image/observation's ID  */
-				subjectId: string,
-				/** The metadata key. If not namespaced, it'll be namespaced to the lightweight protocol's id */
-				key: RemoveNamespace<keyof typeof lightProtocol.metadata> | (string & {}),
-				/** The new value to set it as. Use null to remove the value  */
-				value: null | RuntimeValue | { confidence: number; value?: RuntimeValue }
-			): Promise<void>;
-		};
-	};
-	modals: {
-		byKey(key: `modal_${string}`): Locator;
-		byTitle(message: string): Locator;
-		/**
-		 * Confirms deletion if a deletion-confirm modal is open
-		 * @param type text to type before hitting confirm button
-		 */
-		confirmDeletion(key: `modal_${string}`, type?: string): Promise<void>;
-	};
-	toasts: {
-		byMessage(type: Toast<unknown>['type'] | null, message: string | RegExp): Locator;
-		byType(type: Toast<unknown>['type']): Locator;
-	};
-	settings: {
-		set(values: Partial<Settings>): Promise<void>;
-		get(): Promise<Settings>;
-		get<Key extends keyof Settings>(key: Key): Promise<Settings[Key]>;
-		open(options?: Parameters<Locator['click']>[0]): Promise<void>;
-	};
-	tabs: {
-		go(tab: NavigationTab, options?: Parameters<typeof goToTab>[2]): Promise<void>;
-		get(tab: NavigationTab): Locator;
-	};
-	path: {
-		wait(route: Parameters<typeof waitForRoute>[1]): Promise<void>;
-		go(path: import('$app/types').ResolvedPathname): Promise<void>;
-	};
-	loading: {
-		wait(timeout?: number): Promise<void>;
-		/** Don't wait for loading text to appear, only wait for it to disappear */
-		maybeWait(timeout?: number): Promise<void>;
-		waitIn(area: Locator, timeout?: number): Promise<void>;
-	};
-	sidepanel: Locator;
-	gallery: {
-		area: Locator;
-		/** Get a gallery card in the observations area. */
-		card: (specifier: GalleryCardSpecifier) => Locator;
-		/** Get all cards in the observations area. */
-		cards: () => AsyncIterable<Locator>;
-		/** Select cards. Clears any previous selection. Use continueSelecting to select without clearing */
-		select: (...specifiers: GalleryCardSpecifier[]) => Promise<void>;
-		/** Add to selected cards. Previously selected cards remain selected */
-		selectMore: (...specifiers: GalleryCardSpecifier[]) => Promise<void>;
-	};
-};
 
 const _test = base.extend<
 	{
@@ -212,483 +59,10 @@ const _test = base.extend<
 		{ auto: true },
 	],
 	onnxmodels,
-	app: async ({ page }, use) => {
-		await use({
-			async wait(duration) {
-				await page.waitForTimeout(typeof duration === 'number' ? duration : ms(duration));
-			},
-			sidepanel: page.getByTestId('sidepanel'),
-			metadata: metadataSections(page),
-			db: {
-				async ready() {
-					await page.waitForFunction(() =>
-						Boolean(window.devalue && window.DB && window.refreshDB)
-					);
-				},
-				async refresh() {
-					await page.evaluate(async () => {
-						window.refreshDB();
-					});
-				},
-				async list<T extends keyof IDBDatabaseType>(name: T) {
-					return page.evaluate(async (name) => window.DB.getAll(name), name);
-				},
-				async get<T extends keyof IDBDatabaseType>(name: T, key: string) {
-					return getDatabaseRowById(page, name, key);
-				},
-				async count(name) {
-					return page.evaluate(async (name) => window.DB.count(name), name);
-				},
-				protocol: {
-					byId: async (id) => getDatabaseRowById(page, 'Protocol', id),
-					byName: async (name) => getDatabaseRowByField(page, 'Protocol', 'name', name),
-				},
-				observation: {
-					list: async () => listTable(page, 'Observation'),
-					byId: async (id) => getDatabaseRowById(page, 'Observation', id),
-					byLabel: async (label) =>
-						getDatabaseRowByField(page, 'Observation', 'label', label),
-				},
-				image: {
-					list: async () => listTable(page, 'Image'),
-					byId: async (id) => getDatabaseRowById(page, 'Image', id),
-					byFilename: async (fname) =>
-						getDatabaseRowByField(page, 'Image', 'filename', fname),
-				},
-				session: {
-					list: async () => listTable(page, 'Session'),
-					byId: async (id) => getDatabaseRowById(page, 'Session', id),
-					byName: async (name) => getDatabaseRowByField(page, 'Session', 'name', name),
-				},
-				metadata: {
-					get: async (id) => getDatabaseRowById(page, 'Metadata', id),
-					async of({
-						session,
-						image,
-						imageId,
-						observation: obs,
-						protocolId = lightProtocol.id,
-					}) {
-						let object:
-							| undefined
-							| IDBDatabaseType['Image' | 'Observation' | 'Session']['value'];
-
-						if (imageId) {
-							object = await getDatabaseRowById(page, 'Image', imageId);
-						} else if (image) {
-							object = await getDatabaseRowByField(page, 'Image', 'filename', image);
-						} else if (obs) {
-							object = await getDatabaseRowByField(page, 'Observation', 'label', obs);
-						} else if (session) {
-							object = await getDatabaseRowByField(page, 'Session', 'name', session);
-						} else {
-							throw new Error(
-								'At least one of image, observation or session must be provided'
-							);
-						}
-
-						if (!object) {
-							throw new Error(
-								`Could not find database object for provided parameters: ${JSON.stringify(
-									{ session, image, observation: obs, imageId }
-								)}`
-							);
-						}
-
-						const values =
-							'metadataOverrides' in object
-								? object.metadataOverrides
-								: object.metadata;
-
-						return Object.fromEntries(
-							Object.entries(values)
-								.filter(([id]) =>
-									protocolId ? id.startsWith(`${protocolId}__`) : true
-								)
-								.map(([id, { value, ...rest }]) => [
-									protocolId ? id.replace(`${protocolId}__`, '') : id,
-									{ ...rest, rawValue: value, parsedValue: safeJSONParse(value) },
-								])
-						);
-					},
-					async values(params) {
-						const vals = await this.of(params);
-						return Object.fromEntries(
-							Object.entries(vals).map(([key, { parsedValue }]) => [key, parsedValue])
-						);
-					},
-					async set(subjectId, key, value) {
-						const original =
-							(await getDatabaseRowById(page, 'Image', subjectId)) ??
-							(await getDatabaseRowById(page, 'Observation', subjectId));
-						if (!original) {
-							throw new Error(
-								`Could not find image or observation with ID ${imageId}`
-							);
-						}
-
-						const isObservation = 'metadataOverrides' in original;
-						const subjectType = isObservation ? 'Observation' : 'Image';
-						const property = isObservation ? 'metadataOverrides' : 'metadata';
-
-						if (!key.includes('__')) {
-							key = `${lightProtocol.id}__${key}`;
-						}
-
-						if (value === null) {
-							await page.evaluate(
-								async ([subjectType, subjectId, property, key]) => {
-									const subject = await window.DB.get(subjectType, subjectId);
-									if (!subject) {
-										throw new Error(
-											`Could not find ${subjectType} with ID ${subjectId}`
-										);
-									}
-
-									delete subject[property][key];
-
-									await window.DB.put(subjectType, subject);
-								},
-								[subjectType, subjectId, property, key]
-							);
-							return;
-						}
-
-						const newValue =
-							typeof value === 'object' && 'value' in value
-								? {
-										confidence: value.confidence,
-										value: JSON.stringify(value.value),
-									}
-								: typeof value === 'object' && 'confidence' in value
-									? {
-											confidence: value.confidence,
-											value: original.metadata[key]?.value,
-										}
-									: { confidence: 1, value: JSON.stringify(value) };
-
-						const updated: IDBDatabaseType['Image' | 'Observation']['value'] = {
-							...original,
-							[property]: {
-								...original[property],
-								[key]: {
-									alternatives: [],
-									confidences: { [newValue.value]: newValue.confidence },
-									...newValue,
-								},
-							},
-						};
-
-						await page.evaluate(
-							async ([subjectType, updated]) => {
-								await window.DB.put(subjectType, updated);
-							},
-							[subjectType, updated]
-						);
-					},
-				},
-			},
-			modals: {
-				byKey: (key) => modal(page, { key }),
-				byTitle: (message) => modal(page, { title: message }),
-				confirmDeletion: async (key, type) =>
-					confirmDeletionModal(page, { type, modalKey: key }),
-			},
-			toasts: {
-				byMessage: (type, message) => toast(page, message, { type: type ?? undefined }),
-				byType: (type) => toast(page, null, { type }),
-			},
-			settings: {
-				open: async () => page.getByTestId('app-nav').getByLabel('Réglages').click(),
-				set: async (values) => setSettings({ page }, values),
-				get: async <Key extends keyof Settings>(...maybeKey: [] | [Key]) => {
-					const settings = await getSettings({ page });
-					const key = maybeKey[0];
-					return key ? settings[key] : settings;
-				},
-			},
-			tabs: {
-				go: async (tab, options) => goToTab(page, tab, options),
-				get: (tab) => getTab(page, tab),
-			},
-			path: {
-				wait: async (route) => waitForRoute(page, route),
-				async go(path) {
-					const fullPath = (process.env.BASE_PATH || '') + path;
-					await page.goto(fullPath);
-				},
-			},
-			loading: {
-				wait: async (timeout) => waitForLoadingEnd(page, timeout),
-				waitIn: async (area, timeout) => waitForLoadingEnd(area, timeout),
-				maybeWait: async (timeout) =>
-					waitForLoadingEnd(page, {
-						begin: DontWait,
-						...(timeout ? { finish: timeout } : {}),
-					}),
-			},
-			gallery: {
-				area: page.getByTestId('observations-area'),
-				card: (specifier) => galleryCard(page, specifier),
-				async *cards() {
-					const cardsCount = await galleryCard(page, null).count();
-
-					for (let i = 0; i < cardsCount; i++) {
-						yield galleryCard(page, '#' + i);
-					}
-				},
-				async selectMore(...specifiers) {
-					await page.keyboard.down('ControlOrMeta');
-					for (const specifier of specifiers) {
-						await this.card(specifier).click();
-					}
-					await page.keyboard.up('ControlOrMeta');
-				},
-				async select(...specifiers) {
-					const [first, ...rest] = specifiers;
-					await this.card(first).click();
-					await this.selectMore(...rest);
-				},
-			},
-		});
-	},
-	forEachWorker: [
-		// oxlint-disable-next-line no-empty-pattern required by playwright
-		async ({}, use) => {
-			arthropodaClassifierModel = await getPredownloadedModel(
-				'classifier-collembola.onnx',
-				'classifier-collembola-classmapping.txt'
-			);
-			collembolaClassifierModel = await getPredownloadedModel(
-				'classifier-arthropoda.onnx',
-				'classifier-arthropoda-classmapping.txt'
-			);
-			arthropodaDetectionModel = await getPredownloadedModel('detector-arthropoda.onnx');
-			await use();
-		},
-		{ scope: 'worker', auto: true },
-	],
-	forEachTest: [
-		async ({ page, context, app }, use, info) => {
-			await context.route('**/*', async (route) => {
-				const request = route.request();
-				if (request.resourceType() !== 'document') {
-					return route.fallback();
-				}
-
-				const url = new URL(request.url());
-				const isLocalhost =
-					url.hostname === 'localhost' ||
-					url.hostname === '127.0.0.1' ||
-					url.hostname === '::1';
-				if (!isLocalhost) {
-					return route.fallback();
-				}
-
-				const response = await route.fetch();
-				const headers = {
-					...response.headers(),
-					'Document-Policy': 'js-profiling',
-				};
-
-				await route.fulfill({ response, headers });
-			});
-
-			if (process.env.DEBUG_WORKERS) {
-				let wwcount = 0;
-				page.on('worker', (worker) => {
-					console.debug(`Created WebWorker n°${wwcount + 1} with ${worker.url()}`);
-					wwcount++;
-
-					worker.on('close', () => {
-						console.debug(`Closing WebWorker n°${wwcount}`);
-						wwcount--;
-					});
-				});
-			}
-
-			await rm('./tests/results', { recursive: true, force: true });
-			await mkdir('./tests/results', { recursive: true });
-
-			// XXX: 404 all other protocols so that no other built-in protocols appears before the Example: one in the protocols list
-			// TODO: modify newSession/etc in the tests to target lightProtocol by default instead, test failures happened on main and I was doing sth else
-			const ALLOWED_PROTOCOLS = [fullProtocol.source, lightProtocol.source];
-			await mockUrl(
-				page,
-				context,
-				(u) =>
-					Boolean(
-						new URLPattern('https://*/**/*.cigaleprotocol.*').test(u) &&
-						!ALLOWED_PROTOCOLS.some((source) => {
-							if (typeof source !== 'string') return false;
-							const src = new URL(source);
-							return new URLPattern({
-								hostname: src.hostname,
-								pathname: src.pathname.replace(
-									'cigaleapp/cigale/main/',
-									'cigaleapp/cigale/:branch/'
-								),
-							}).test(u);
-						})
-					),
-				{
-					status: 404,
-					body: '',
-				}
-			);
-
-			if (!info.tags.includes('@real-protocol')) {
-				// @ts-expect-error we don't support non-string protocol source values for now
-				await mockProtocolSourceURL(page, context, fullProtocol.source, {
-					json: lightProtocol,
-				});
-
-				await mockProtocolSourceURL(page, context, lightProtocol.source, {
-					json: lightProtocol,
-				});
-			}
-
-			await mockPredownloadedModels(page, context, fullProtocol, {
-				crop: [arthropodaDetectionModel],
-				species: [collembolaClassifierModel, arthropodaClassifierModel],
-			});
-
-			const concurrency = info.annotations.find((a) => a.type === 'concurrency')?.description;
-			if (concurrency) {
-				await setHardwareConcurrency(page, Number.parseInt(concurrency));
-			}
-
-			if (
-				info.tags.includes('@webkit-no-parallelization') &&
-				context.browser()?.browserType().name() === 'webkit'
-			) {
-				await setHardwareConcurrency(page, 1);
-			}
-
-			if (!info.tags.includes('@blank')) {
-				await page.goto('./');
-				await app.db.ready();
-			}
-
-			await use();
-
-			if (info.status !== info.expectedStatus) {
-				info.attach('database snapshot', {
-					body: await dumpDatabase(page, null),
-				});
-
-				info.attach('UI state', {
-					body: await page.evaluate(() =>
-						JSON.stringify(window.uiState?.snapshot() ?? null, null, 2)
-					),
-				});
-			}
-		},
-		{ auto: true },
-	],
+	app,
+	forEachWorker: [forEachWorker, { scope: 'worker', auto: true }],
+	forEachTest: [forEachTest, { auto: true }],
 });
-
-export const assert = baseExpect.extend({
-	async toHaveTooltip(
-		locator: Locator,
-		/** null to expect no tooltip on the element */
-		expected: string | RegExp | null,
-		options?: { timeout?: number }
-	) {
-		const assertionName = 'toHaveTooltip';
-
-		/**
-		 * Used when expected=null and we failed (we got a tooltip)
-		 * since in that case, expectTooltipContent() does not throw
-		 *
-		 * We do it like this instead of handling it in expectTooltipContent,
-		 * so that we still get the timeout behavior (ie checking that there's really no tooltip,
-		 * even if it takes a little while to show up)
-		 */
-		let tooltipTextActual = '';
-
-		let pass: boolean;
-		// eslint-disable-next-line @typescript-eslint/no-explicit-any
-		let matcherResult: any;
-		try {
-			tooltipTextActual = await expectTooltipContent(
-				locator.page(),
-				locator,
-				expected === null ? /.*/ : expected,
-				options
-			);
-			pass = true;
-			// eslint-disable-next-line @typescript-eslint/no-explicit-any
-		} catch (e: any) {
-			matcherResult = e.matcherResult;
-			pass = false;
-		}
-
-		if (expected === null) pass = !pass;
-		if (this.isNot) pass = !pass;
-
-		const message = () =>
-			this.utils.matcherHint(assertionName, undefined, undefined, {
-				isNot: this.isNot,
-			}) +
-			'\n\n' +
-			`Locator: ${locator}\n` +
-			`Expected: ${this.isNot ? 'not' : ''} ${expected === null ? 'no tooltip' : `tooltip with ${this.utils.printExpected(expected)}`}\n` +
-			(matcherResult || tooltipTextActual
-				? `Received: ${this.utils.printReceived(tooltipTextActual || matcherResult.actual)}`
-				: '');
-
-		return {
-			message,
-			pass,
-			name: assertionName,
-			expected,
-			actual: matcherResult?.actual,
-		};
-	},
-
-	async toBeOnSlide(locator: Locator, slideName: string, options?: { timeout?: number }) {
-		const assertionName = 'toBeOnSlide';
-
-		let pass: boolean;
-		// eslint-disable-next-line @typescript-eslint/no-explicit-any
-		let matcherResult: any;
-
-		try {
-			const currentSlide = locator.locator("[data-current-slide='true']");
-			await assert(currentSlide).toHaveAccessibleName(slideName, options);
-			pass = true;
-			// eslint-disable-next-line @typescript-eslint/no-explicit-any
-		} catch (e: any) {
-			matcherResult = e.matcherResult;
-			pass = false;
-		}
-
-		if (this.isNot) pass = !pass;
-
-		const message = () =>
-			this.utils.matcherHint(assertionName, undefined, undefined, {
-				isNot: this.isNot,
-			}) +
-			'\n\n' +
-			`Locator: ${locator}\n` +
-			`Expected: ${this.isNot ? 'not' : ''} to be on slide ${this.utils.printExpected(slideName)}\n` +
-			(matcherResult
-				? `Received: is on slide ${this.utils.printReceived(matcherResult.actual)}`
-				: '');
-
-		return {
-			message,
-			pass,
-			name: assertionName,
-			expected: slideName,
-			actual: matcherResult?.actual,
-		};
-	},
-});
-
-// Encourage using soft expects in tests, and only use hard expects (assert in our case) when it's necessary for the rest of the test to continue
-export const expect = assert.soft;
 
 /**
  * Default test has a "empty" storage state, that has the protocol already imported but no sessions
